Refactoring Storybook stories.
