+++ app/javascript/articles/__stories__/LoadingArticle.stories.jsx
@@ -1,19 +1,9 @@
 import { h } from 'preact';
 import { storiesOf } from '@storybook/react';
 import { LoadingArticle } from '..';
-import { defaultChildrenPropTypes } from '../../src/components/common-prop-types/default-children-prop-types';
 import '../../../assets/stylesheets/articles.scss';
+import { articleDecorator } from './articleDecorator';
 
-const ArticleWrapper = ({ children }) => (
-  <div className="articles-list">{children}</div>
-);
-ArticleWrapper.propTypes = {
-  children: defaultChildrenPropTypes.isRequired,
-};
-ArticleWrapper.displayName = 'ArticleWrapper';
-
-storiesOf('Components/Articles/Loading Article', module).add('Default', () => (
-  <ArticleWrapper>
-    <LoadingArticle />
-  </ArticleWrapper>
-));
+storiesOf('Components/Articles', module)
+  .addDecorator(articleDecorator)
+  .add('Article Loading', () => <LoadingArticle />);
